refactor(server): move server wiring into WebSocketServer
- reduce server entrypoint to composition-only while preserving
  the existing http and websocket behavior
